Use new AdminTable component for ManageChallenge
